Running docker container as root instead of user shiny
attempting to solve https://github.com/rocker-org/shiny/issues/79. There
are some suggested approaches to run as non-root user, but I am not sure
about all the details.

Running as root for now to fix the issue. The official docs does not
mention anything so I would rather wait for an update so that we run in
a recommended way without too many workarounds